add create fabric garment category and measurement

+++ database/migrations/2026_04_12_131436_create_garments_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('name');
             $table->text('description')->nullable();
             $table->string('slug');
-            $table->float('base_price');
+            $table->decimal('base_price', 10, 2);
             $table->string('image');
             $table->timestamps();
         });
